basic practice feature done
The user can practice the words of a specific vocbook by filling in the korean word into a textbox. Afterward, it is shown to the user if his answer is correct or not. If not, the correct answer is shown. Clicking on the 'next word' button repeats this process until no word is left. Then a table is shown to the user with the answers he gave.  By clicking on the button 'practice again' the user can start a new practice round. The answers of the previous round are deleted.