changed structure of index page and how the books are displayed. Also added more lessons.
The user  starts initially with only one lesson in the desk area. The user can add more lessons to the desk. The user can move books from the desk to another area,, the drawer. Here are all books that the user is tired of learning or maybe has already 100% done. If the user likes to learn one of the books in the drawer area in the future, the user can move it back to the desk area. This split is just an organizational adjustment so the user is not overwhelmed with too many lessons from the start.